fix: rewrite content-disposition in gazette repair script too
The repair script re-publishes a gazette's S3 object (strips the
scheduledAt tag) but never rewrote its Content-Disposition, so a
repaired gazette would still download under its raw S3 key instead of
its title — inconsistent with the ingestion cron, which now does this
on every publish.

Moved getContentDispositionForTitle from asset.service.ts into
@isomer/algolia (already the shared home for cron/repair gazette
logic, e.g. buildGazetteSearchRecords) so both the cron and the repair
script use the same implementation instead of duplicating it. The
repair script mirrors setAssetAsPublished's self-copy-with-
MetadataDirective-REPLACE approach, since S3 object metadata is
immutable.
