+++ apps/studio/src/server/cron/jobs/schedulePushDocumentJob.ts
@@ -6,7 +6,6 @@ import { ENABLE_SEARCHSG_GAZETTE_INGESTION } from "~/lib/growthbook"
 import { createBaseLogger } from "~/lib/logger"
 import { getBlob, setAssetAsPublished } from "~/lib/s3"
 import { createGrowthBookContext } from "~/server/context"
-import { getContentDispositionForTitle } from "~/server/modules/asset/asset.service"
 import { db } from "~/server/modules/database"
 import {
   generateDocumentId,
@@ -15,6 +14,7 @@ import {
 
 import {
   buildGazetteSearchRecords,
+  getContentDispositionForTitle,
   parseFullTextFromPDF,
 } from "@isomer/algolia"
 import { registerPgbossJob } from "@isomer/pgboss"
